+++ unittest/heap_test.cc
@@ -34,7 +34,7 @@ class HeapTest : public testing::Test {
   virtual ~HeapTest();
   // Pushes the test data onto both the heap and the KDVector.
   void PushTestData(GenericHeap<IntKDPair>* heap, KDVector* v) {
-    for (int i = 0; i < ARRAYSIZE(test_data); ++i) {
+    for (size_t i = 0; i < ARRAYSIZE(test_data); ++i) {
       IntKDPair pair(test_data[i], i);
       heap->Push(&pair);
       v->push_back(pair);
@@ -137,7 +137,7 @@ TEST_F(HeapTest, RevalueTest) {
   GenericHeap<PtrPair> heap;
   GenericVector<PtrPair> v;
   // Push the test data onto both the heap and the vector.
-  for (int i = 0; i < ARRAYSIZE(test_data); ++i) {
+  for (size_t i = 0; i < ARRAYSIZE(test_data); ++i) {
     PtrPair h_pair;
     h_pair.key = test_data[i];
     PtrPair v_pair;
